[style] edit event list
