--- src/container/Event/components/DateEvent/index.style.ts
@@ -1,12 +1,0 @@
-import styled from '@emotion/styled';
-
-export const DateEventTextStyled = styled.span`
-  font-color: var(--tg-theme-hint-color);
-  font-size: 0.875rem;
-  font-weight: bold;
-`;
-
-export const DateEventStyled = styled.span`
-  font-color: var(--tg-theme-hint-color);
-  font-size: 0.875rem;
-`;
