--- src/container/Event/components/DateEvent/types.ts
@@ -1,3 +1,0 @@
-export interface DateEventProps {
-  date: number | string | Date;
-}
