--- src/container/Event/components/Description/index.style.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const DescriptionStyled = styled.span`
-  font-color: var(--tg-theme-text-color);
-  font-size: 1rem;
-  margin: 8px 0;
-`;
