--- src/container/Event/components/Description/types.ts
@@ -1,3 +1,0 @@
-export interface DescriptionProps {
-  text: string;
-}
