--- src/container/Event/components/Title/index.style.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const TitleStyled = styled.span`
-  font-color: var(--tg-theme-text-color);
-  font-size: 1.25rem;
-  font-weight: bold;
-`;
