--- src/container/Event/components/Title/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { TitleProps } from './types';
-import { TitleStyled } from './index.style';
-
-const Title = ({ text }: TitleProps): React.ReactElement => {
-  return <TitleStyled>{text}</TitleStyled>;
-};
-
-export default Title;
