--- src/container/Event/components/Title/types.ts
@@ -1,3 +1,0 @@
-export interface TitleProps {
-  text: string;
-}
