+++ src/container/EventList/components/Event/components/Button/index.style.ts
@@ -4,5 +4,4 @@ import Button from '@mui/material/Button';
 export const ButtonStyled = styled(Button)`
   background-color: var(--tg-theme-button-color);
   color: var(--tg-theme-button-text-color);
-  margin-top: 16px;
 `;
