+++ src/container/EventList/components/Event/components/Description/index.style.ts
@@ -3,5 +3,5 @@ import styled from '@emotion/styled';
 export const DescriptionStyled = styled.span`
   color: var(--tg-theme-text-color);
   font-size: 1rem;
-  margin-top: 24px;
+  margin-bottom: 24px;
 `;
