+++ src/container/EventList/components/Event/components/Title/index.style.ts
@@ -4,4 +4,5 @@ export const TitleStyled = styled.span`
   color: var(--tg-theme-text-color);
   font-size: 1.25rem;
   font-weight: bold;
+  margin-bottom: 24px;
 `;
