+++ src/container/EventList/components/Event/index.tsx
@@ -4,7 +4,12 @@ import Title from './components/Title';
 import Description from './components/Description';
 import DateEvent from './components/DateEvent';
 import Button from './components/Button';
-import { EventProps } from './types';
+
+interface EventProps {
+  name: string;
+  description: string;
+  date: number | string | Date;
+}
 
 const Event = ({ name, description, date }: EventProps): React.ReactElement => {
   const handleParticipateEvent = async (event: React.FormEvent<HTMLFormElement>) => {
