--- src/container/EventList/components/Event/types.ts
@@ -1,5 +1,0 @@
-export interface EventProps {
-  name: string;
-  description: string;
-  date: number | string | Date;
-}
