+++ src/container/EventList/index.style.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
 
 export const EventListStyled = styled.div`
-  margin-top: 32px;
+  margin-top: 8px;
 `;
